QIC model/requirements.txt
torch==2.2.1
transformers==4.41.1
datasets==2.19.0
scikit-learn==1.4.2
evaluate==0.4.2
accelerate==0.30.0
pandas==2.2.2
tqdm==4.66.4
pyyaml==6.0.1
joblib==1.4.2
numpy==1.26.4 